+++ app/src/main/java/com/preonboarding/locationhistory/data/RoomMapper.kt
@@ -1,6 +1,6 @@
 package com.preonboarding.locationhistory.data
 
-import com.preonboarding.locationhistory.data.local.entity.LocationEntity
+import com.preonboarding.locationhistory.data.db.entity.LocationEntity
 import com.preonboarding.locationhistory.domain.model.Location
 
 fun Location.toEntity() = LocationEntity(
