+++ app/src/main/java/com/preonboarding/locationhistory/data/db/dao/LocationDao.kt
@@ -1,10 +1,10 @@
-package com.preonboarding.locationhistory.data.local.dao
+package com.preonboarding.locationhistory.data.db.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.preonboarding.locationhistory.data.local.entity.LocationEntity
+import com.preonboarding.locationhistory.data.db.entity.LocationEntity
 
 @Dao
 interface LocationDao {
