+++ app/src/main/java/com/preonboarding/locationhistory/data/db/database/LocationDatabase.kt
@@ -1,9 +1,9 @@
-package com.preonboarding.locationhistory.data.local.database
+package com.preonboarding.locationhistory.data.db.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.preonboarding.locationhistory.data.local.dao.LocationDao
-import com.preonboarding.locationhistory.data.local.entity.LocationEntity
+import com.preonboarding.locationhistory.data.db.dao.LocationDao
+import com.preonboarding.locationhistory.data.db.entity.LocationEntity
 
 @Database(entities = [LocationEntity::class], version = 1, exportSchema = false)
 abstract class LocationDatabase: RoomDatabase() {
